changing navigation will cancel timer new add and minus operators
